aws/requirements.txt
boto3>=1.26.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.20.0
tavily
